+++ client.go
@@ -15,6 +15,12 @@ var defaultClientTimeout = 60 * time.Second
 
 // this is a client for a ReductStore instance.
 type Client interface {
+	// Get Info
+	GetInfo(ctx context.Context) (model.ServerInfo, error)
+	// Check if the storage engine is working
+	IsLive(ctx context.Context) (bool, error)
+	// Get a list of the buckets with their stats
+	GetBuckets(ctx context.Context) ([]model.BucketInfo, error)
 	// Create a new bucket
 	CreateBucket(ctx context.Context, name string, settings *model.BucketSetting) (Bucket, error)
 	// Create a new bucket if it doesn't exist and return it
@@ -58,6 +64,36 @@ func NewClient(url string, options ClientOptions) Client {
 	return client
 }
 
+// GetInfo returns information about the server.
+func (c *ReductClient) GetInfo(ctx context.Context) (model.ServerInfo, error) {
+	var info model.ServerInfo
+	err := c.HTTPClient.Get(ctx, "/info", &info)
+	if err != nil {
+		return model.ServerInfo{}, model.APIError{Message: err.Error(), Original: err}
+	}
+	return info, nil
+}
+
+// IsLive checks if the server is live.
+func (c *ReductClient) IsLive(ctx context.Context) (bool, error) {
+	err := c.HTTPClient.Head(ctx, "/alive")
+	if err != nil {
+		return false, model.APIError{Message: err.Error(), Original: err}
+	}
+	return true, nil
+}
+
+// GetBuckets returns a list of buckets with their stats.
+func (c *ReductClient) GetBuckets(ctx context.Context) ([]model.BucketInfo, error) {
+	var buckets map[string][]model.BucketInfo
+	err := c.HTTPClient.Get(ctx, "/list", &buckets)
+	if err != nil {
+		return nil, model.APIError{Message: err.Error(), Original: err}
+	}
+	return buckets["buckets"], nil
+}
+
+// GetBucket returns a bucket.
 func (c *ReductClient) GetBucket(ctx context.Context, name string) (Bucket, error) {
 	err := c.HTTPClient.Get(ctx, fmt.Sprintf(`/b/%s`, name), nil)
 	if err != nil {
@@ -90,6 +126,7 @@ func (c *ReductClient) CreateOrGetBucket(ctx context.Context, name string, setti
 	return NewBucket(name, c.HTTPClient), err
 }
 
+// CheckBucketExists checks if a bucket exists.
 func (c *ReductClient) CheckBucketExists(ctx context.Context, name string) (bool, error) {
 	err := c.HTTPClient.Head(ctx, fmt.Sprintf(`/b/%s`, name))
 	if err != nil {
@@ -98,6 +135,7 @@ func (c *ReductClient) CheckBucketExists(ctx context.Context, name string) (bool
 	return true, nil
 }
 
+// RemoveBucket removes a bucket.
 func (c *ReductClient) RemoveBucket(ctx context.Context, name string) error {
 	return c.HTTPClient.Delete(ctx, fmt.Sprintf(`/b/%s`, name))
 }