+++ client_test.go
@@ -19,6 +19,24 @@ func TestCreateOrGetBucket_Success(t *testing.T) {
 	assert.Equal(t, bucket.Name, mainTestBucket.Name)
 }
 
+func TestGetInfo(t *testing.T) {
+	info, err := client.GetInfo(context.Background())
+	assert.NoError(t, err)
+	assert.NotNil(t, info)
+}
+
+func TestGetBuckets(t *testing.T) {
+	buckets, err := client.GetBuckets(context.Background())
+	assert.NoError(t, err)
+	assert.NotNil(t, buckets)
+}
+
+func TestHealth(t *testing.T) {
+	isLive, err := client.IsLive(context.Background())
+	assert.NoError(t, err)
+	assert.True(t, isLive)
+}
+
 // Creating a new bucket should succeed.
 func TestCreateBucket_Success(t *testing.T) {
 	ctx := context.Background()